[fix]: [CCM-9902]: Handle Cost Category delete request (#40360)
* [fix]: [CCM-9902]: Handle Cost Category delete request

* [fix]: [CCM-9902]: Handle Cost Category delete request

* [fix]: [CCM-9902]: Handle Cost Category delete request

* [fix]: [CCM-9902]: Handle Cost Category delete request

* [fix]: [CCM-9902]: Handle Cost Category delete request

* [fix]: [CCM-9902]: Adding new list API here to find related perspectives to BM

* [fix]: [CCM-9902]: Implementing new list API

* [fix]: [CCM-9902]: Adding more info to API

* [fix]: [CCM-9902]: Making getPerspectives API POST API

* [fix]: [CCM-9902]: Fixing * import

* [fix]: [CCM-9902]: Adding some log lines

* [fix]: [CCM-9902]: Reverting the log lines d089fbe

* [fix]: [CCM-9902]: Fixing test cases

* [fix]: [CCM-9902]: Including review comments

* [fix]: [CCM-9902]: Including more review comments

* [fix]: [CCM-9902]: Fix test case

* [fix]: [CCM-9902]: Removing unused imports

+++ ce-nextgen/service/src/main/java/io/harness/ccm/remote/resources/BusinessMappingResource.java
@@ -19,6 +19,8 @@ import io.harness.ccm.audittrails.events.CostCategoryUpdateEvent;
 import io.harness.ccm.rbac.CCMRbacHelper;
 import io.harness.ccm.views.businessMapping.entities.BusinessMapping;
 import io.harness.ccm.views.businessMapping.service.intf.BusinessMappingService;
+import io.harness.ccm.views.dto.CostCategoryDeleteDTO;
+import io.harness.ccm.views.dto.LinkedPerspectives;
 import io.harness.ccm.views.service.CEViewService;
 import io.harness.exception.InvalidRequestException;
 import io.harness.outbox.api.OutboxService;
@@ -29,9 +31,18 @@ import com.codahale.metrics.annotation.ExceptionMetered;
 import com.codahale.metrics.annotation.Timed;
 import com.google.inject.Inject;
 import com.google.inject.name.Named;
+import io.fabric8.utils.Lists;
+import io.fabric8.utils.Maps;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import io.swagger.v3.oas.annotations.Hidden;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.parameters.RequestBody;
+import java.util.Collections;
 import java.util.List;
+import javax.validation.Valid;
+import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
@@ -40,6 +51,7 @@ import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
+import javax.ws.rs.core.MediaType;
 import lombok.extern.slf4j.Slf4j;
 import net.jodah.failsafe.Failsafe;
 import net.jodah.failsafe.RetryPolicy;
@@ -62,6 +74,8 @@ public class BusinessMappingResource {
   @Inject CEViewService ceViewService;
 
   private final RetryPolicy<Object> transactionRetryPolicy = DEFAULT_RETRY_POLICY;
+  private static final String COST_CATEGORY_NOT_DELETED = "Cost Category exists in listed perspective(s)";
+  private static final String COST_CATEGORY_DELETED = "Cost category is safely deleted";
 
   @POST
   @Timed
@@ -136,15 +150,57 @@ public class BusinessMappingResource {
   @Timed
   @ExceptionMetered
   @ApiOperation(value = "Delete Business Mapping", nickname = "deleteBusinessMapping")
-  public RestResponse<String> delete(@NotEmpty @QueryParam(NGCommonEntityConstants.ACCOUNT_KEY) String accountId,
-      @PathParam("id") String businessMappingId) {
+  public RestResponse<CostCategoryDeleteDTO> delete(@NotEmpty @QueryParam(NGCommonEntityConstants.ACCOUNT_KEY)
+                                                    String accountId, @PathParam("id") String businessMappingId) {
     rbacHelper.checkCostCategoryDeletePermission(accountId, null, null);
-    BusinessMapping costCategory = businessMappingService.get(businessMappingId, accountId);
-    businessMappingService.delete(businessMappingId, accountId);
-    Failsafe.with(transactionRetryPolicy).get(() -> transactionTemplate.execute(status -> {
-      outboxService.save(new CostCategoryDeleteEvent(accountId, costCategory.toDTO()));
-      return true;
-    }));
-    return new RestResponse<>("Successfully deleted the Business Mapping");
+    LinkedPerspectives perspectiveListMessage =
+        ceViewService.getViewsByBusinessMapping(accountId, Collections.singletonList(businessMappingId)).get(0);
+    CostCategoryDeleteDTO.CostCategoryDeleteDTOBuilder costCategoryDeleteDTOBuilder =
+        CostCategoryDeleteDTO.builder()
+            .linkedPerspectives(perspectiveListMessage.getPerspectiveIdAndName())
+            .deleted(false)
+            .message(COST_CATEGORY_NOT_DELETED);
+
+    // No dependency on cost category(business mapping) exists
+    if (Maps.isNullOrEmpty(perspectiveListMessage.getPerspectiveIdAndName())) {
+      BusinessMapping costCategory = businessMappingService.get(businessMappingId, accountId);
+      businessMappingService.delete(businessMappingId, accountId);
+      costCategoryDeleteDTOBuilder.deleted(true).message(COST_CATEGORY_DELETED);
+      Failsafe.with(transactionRetryPolicy).get(() -> transactionTemplate.execute(status -> {
+        outboxService.save(new CostCategoryDeleteEvent(accountId, costCategory.toDTO()));
+        return true;
+      }));
+    }
+    return new RestResponse<>(costCategoryDeleteDTOBuilder.build());
+  }
+
+  @POST
+  @Path("linkedPerspectives")
+  @Timed
+  @Hidden
+  @Consumes(MediaType.APPLICATION_JSON)
+  @ExceptionMetered
+  @ApiOperation(value = "Get related Perspectives given Business Mapping Ids",
+      nickname = "getPerspectivesGivenBusinessMappingIds")
+  @Operation(operationId = "getPerspectivesGivenBusinessMappingIds",
+      description =
+          "Given Business Mapping Id, find out Perspectives where Business Mapping is present in Rules or Group By",
+      summary = "Return list of perspectives corresponding to Business Mapping Id",
+      responses =
+      {
+        @io.swagger.v3.oas.annotations.responses.
+        ApiResponse(description = "Return list of perspectives corresponding to Business Mapping Id",
+            content = { @Content(mediaType = MediaType.APPLICATION_JSON) })
+      })
+  public RestResponse<List<LinkedPerspectives>>
+  getPerspectives(@QueryParam(NGCommonEntityConstants.ACCOUNT_KEY) String accountId,
+      @RequestBody(
+          required = true, description = "List of Business Mapping Ids") @Valid List<String> businessMappingIds) {
+    rbacHelper.checkCostCategoryViewPermission(accountId, null, null);
+    List<LinkedPerspectives> perspectiveListMessages = null;
+    if (!Lists.isNullOrEmpty(businessMappingIds)) {
+      perspectiveListMessages = ceViewService.getViewsByBusinessMapping(accountId, businessMappingIds);
+    }
+    return new RestResponse<>(perspectiveListMessages);
   }
 }
